editor: dedupe self-collision + per-link mesh extraction
Three call sites (webserver._build_collision, _autolimits_cli, the viser
UI) each had their own copy of the skrobot visual_mesh -> {link: trimesh}
extraction loop; fold them into autoinit.link_visual_mesh / link_meshes.

ui.CollisionWatcher was a near-verbatim second implementation of
autoinit.SelfCollision (same rest+adjacency baseline, convex-hull proxy,
_pairs/min_distance). Delete it and the duplicate _collision_mesh /
_link_mesh helpers; the UI now uses autoinit.SelfCollision, with a new
offenders() convenience returning (new_pairs, offending_links).

No behavior change (codex-reviewed); net -70 lines.
